chore: backport changes to catalog from enterprise (#26116)
* chore: backport changes to influxdb3_catalog crate

* chore: backport changes to influxdb3_cache crate

* chore: backport changes to influxdb3_write crate

* chore: backport changes to influxdb3_proc_eng crate

* chore: backport influxdb3 crate changes for catalog

* chore: backport changes to influxdb3_id crate

* chore: backport changes to influxdb3_wal crate

* chore: backport changes to influxdb3_clap_blocks crate

* chore: backport changes to influxdb3_client crate

* chore: backport influxdb3_server crate changes

* chore: fix after full backport

* fix: ordering of catalog broadcast

+++ influxdb3_write/src/write_buffer/persisted_files.rs
@@ -413,7 +413,7 @@ mod tests {
         parquet_files: Vec<ParquetFile>,
         snapshot_id: u64,
         wal_id: u64,
-        catalog_id: u32,
+        catalog_id: u64,
     ) -> PersistedSnapshot {
         let snap1 = SnapshotSequenceNumber::new(snapshot_id);
         let wal1 = WalFileSequenceNumber::new(wal_id);
